refactor: move queries into their own folder
