+++ web-registry/src/lib/queries/react-query/sanity/queries.sanity.ts
@@ -1,0 +1,24 @@
+import {
+  AllCreditClassDocument,
+  AllCreditClassQuery,
+} from 'generated/sanity-graphql';
+
+import {
+  ReactQueryGetAllCreditClassesParams,
+  ReactQueryGetAllCreditClassesResponse,
+} from './queries.sanity.types';
+
+export const getAllCreditClassesQuery = ({
+  sanityClient,
+}: ReactQueryGetAllCreditClassesParams): ReactQueryGetAllCreditClassesResponse => ({
+  queryKey: ['AllCreditClassQuery'],
+  queryFn: async () => {
+    const { data: sanityCreditClassData } =
+      await sanityClient.query<AllCreditClassQuery>({
+        query: AllCreditClassDocument,
+      });
+
+    return sanityCreditClassData;
+  },
+  staleTime: Infinity,
+});
